Align List/Detail pages with PF Improve Navigation active prop detection
