--- src/app/components/AppLayout/AppNavigation.css
@@ -1,12 +1,0 @@
-/**
- Highlighting the active NavItem:
-  - Done by detecting the "active" class set by React Router on the active Route
-  - Highlights the entry for the active Route
-  - Highlights the section that includes the active Route, but only when it's collapsed
- */
-#flightclt-nav .pf-v5-c-nav__item:not(.pf-m-expanded):has(.pf-v5-c-nav__link.active) button {
-  background-color: var(--pf-v5-c-nav__link--active--BackgroundColor);
-}
-#flightclt-nav .pf-v5-c-nav__item:not(.pf-m-expandable):has(.pf-v5-c-nav__link.active) {
-  background-color: var(--pf-v5-c-nav__link--active--BackgroundColor);
-}
