+++ src/app/components/Fleet/FleetDetails/Tabs/FleetDevicesTab.tsx
@@ -1,29 +1,23 @@
 import * as React from 'react';
-import { Alert, EmptyState, EmptyStateBody, Spinner } from '@patternfly/react-core';
+import { Alert, Bullseye, EmptyState, EmptyStateBody, Spinner } from '@patternfly/react-core';
 
 import { getErrorMessage } from '@app/utils/error';
 import { DeviceTable } from '@app/components/Device/DeviceList';
 import { useFetchPeriodically } from '@app/hooks/useFetchPeriodically';
 import { DeviceList } from '@types';
 
-const FleetDevicesTab = ({
-  fleetName,
-  onDevicesLoaded,
-}: {
-  fleetName: string;
-  onDevicesLoaded: (deviceCount: number) => void;
-}) => {
+const FleetDevicesTab = ({ fleetName }: { fleetName: string }) => {
   const [deviceList, loadingDevices, devicesError, refetch] = useFetchPeriodically<DeviceList>({
     endpoint: `devices?owner=Fleet/${fleetName}`,
   });
 
-  React.useEffect(() => {
-    onDevicesLoaded(deviceList?.items?.length || 0);
-  }, [deviceList, onDevicesLoaded]);
-
   let tabContent;
   if (loadingDevices || deviceList === undefined) {
-    tabContent = <Spinner />;
+    tabContent = (
+      <Bullseye>
+        <Spinner />
+      </Bullseye>
+    );
   }
   if (deviceList?.items) {
     tabContent =
