+++ src/app/components/Repository/RepositoryDetails.tsx
@@ -13,8 +13,6 @@ import {
   Grid,
   GridItem,
   Icon,
-  Stack,
-  StackItem,
 } from '@patternfly/react-core';
 import { CheckCircleIcon, ExternalLinkAltIcon, InfoCircleIcon } from '@patternfly/react-icons';
 
@@ -53,92 +51,88 @@ const RepositoryDetails = () => {
       title={repoDetails?.metadata.name}
       resourceLink="/administration/repositories"
       resourceName="Repositories"
+      actions={
+        <DetailsPageActions>
+          <DropdownList>{deleteAction}</DropdownList>
+        </DetailsPageActions>
+      }
     >
-      <Stack hasGutter>
-        <StackItem>
-          <DetailsPageActions>
-            <DropdownList>{deleteAction}</DropdownList>
-          </DetailsPageActions>
-        </StackItem>
-        <StackItem>
-          <Grid hasGutter>
-            <GridItem>
-              <Card>
-                <CardTitle>Details</CardTitle>
-                <CardBody>
-                  <DescriptionList columnModifier={{ lg: '2Col' }}>
-                    <DescriptionListGroup>
-                      <DescriptionListTerm>Url</DescriptionListTerm>
-                      <DescriptionListDescription>
-                        <Button
-                          component="a"
-                          variant="link"
-                          isInline
-                          href={repoDetails?.spec.repo}
-                          target="_blank"
-                          icon={<ExternalLinkAltIcon />}
-                          iconPosition="end"
-                        >
-                          {repoDetails?.spec.repo}
-                        </Button>
-                      </DescriptionListDescription>
-                    </DescriptionListGroup>
-                    <DescriptionListGroup>
-                      <DescriptionListTerm>Status</DescriptionListTerm>
-                      <DescriptionListDescription>
-                        {' '}
-                        {repoDetails ? <StatusInfo statusInfo={getRepositorySyncStatus(repoDetails)} /> : '-'}
-                      </DescriptionListDescription>
-                    </DescriptionListGroup>
-                    <DescriptionListGroup>
-                      <DescriptionListTerm>Last transition</DescriptionListTerm>
-                      <DescriptionListDescription>
-                        {repoDetails ? getRepositoryLastTransitionTime(repoDetails) : '-'}
-                      </DescriptionListDescription>
-                    </DescriptionListGroup>
-                    <DescriptionListGroup>
-                      <DescriptionListTerm>Username</DescriptionListTerm>
-                      <DescriptionListDescription>{repoDetails?.spec.username || '-'}</DescriptionListDescription>
-                    </DescriptionListGroup>
-                    <DescriptionListGroup>
-                      <DescriptionListTerm>Password</DescriptionListTerm>
-                      <DescriptionListDescription>
-                        {repoDetails?.spec.password ? (
-                          <>
-                            <Icon status="success">
-                              <CheckCircleIcon />
-                            </Icon>{' '}
-                            Password is set
-                          </>
-                        ) : (
-                          <>
-                            <Icon status="info">
-                              <InfoCircleIcon />
-                            </Icon>{' '}
-                            Password not set
-                          </>
-                        )}
-                      </DescriptionListDescription>
-                    </DescriptionListGroup>
-                  </DescriptionList>
-                </CardBody>
-              </Card>
-            </GridItem>
-            <GridItem>
-              <Card>
-                <CardTitle>
-                  <WithHelperText popoverContent="Flight control will monitor the specified paths, import the defined fleets and synchronise devices">
-                    Resource syncs
-                  </WithHelperText>
-                </CardTitle>
-                <CardBody>
-                  <RepositoryResourceSyncList repositoryId={repositoryId} />
-                </CardBody>
-              </Card>
-            </GridItem>
-          </Grid>
-        </StackItem>
-      </Stack>
+      <Grid hasGutter>
+        <GridItem>
+          <Card>
+            <CardTitle>Details</CardTitle>
+            <CardBody>
+              <DescriptionList columnModifier={{ lg: '2Col' }}>
+                <DescriptionListGroup>
+                  <DescriptionListTerm>Url</DescriptionListTerm>
+                  <DescriptionListDescription>
+                    <Button
+                      component="a"
+                      variant="link"
+                      isInline
+                      href={repoDetails?.spec.repo}
+                      target="_blank"
+                      icon={<ExternalLinkAltIcon />}
+                      iconPosition="end"
+                    >
+                      {repoDetails?.spec.repo}
+                    </Button>
+                  </DescriptionListDescription>
+                </DescriptionListGroup>
+                <DescriptionListGroup>
+                  <DescriptionListTerm>Status</DescriptionListTerm>
+                  <DescriptionListDescription>
+                    {' '}
+                    {repoDetails ? <StatusInfo statusInfo={getRepositorySyncStatus(repoDetails)} /> : '-'}
+                  </DescriptionListDescription>
+                </DescriptionListGroup>
+                <DescriptionListGroup>
+                  <DescriptionListTerm>Last transition</DescriptionListTerm>
+                  <DescriptionListDescription>
+                    {repoDetails ? getRepositoryLastTransitionTime(repoDetails) : '-'}
+                  </DescriptionListDescription>
+                </DescriptionListGroup>
+                <DescriptionListGroup>
+                  <DescriptionListTerm>Username</DescriptionListTerm>
+                  <DescriptionListDescription>{repoDetails?.spec.username || '-'}</DescriptionListDescription>
+                </DescriptionListGroup>
+                <DescriptionListGroup>
+                  <DescriptionListTerm>Password</DescriptionListTerm>
+                  <DescriptionListDescription>
+                    {repoDetails?.spec.password ? (
+                      <>
+                        <Icon status="success">
+                          <CheckCircleIcon />
+                        </Icon>{' '}
+                        Password is set
+                      </>
+                    ) : (
+                      <>
+                        <Icon status="info">
+                          <InfoCircleIcon />
+                        </Icon>{' '}
+                        Password not set
+                      </>
+                    )}
+                  </DescriptionListDescription>
+                </DescriptionListGroup>
+              </DescriptionList>
+            </CardBody>
+          </Card>
+        </GridItem>
+        <GridItem>
+          <Card>
+            <CardTitle>
+              <WithHelperText popoverContent="Flight control will monitor the specified paths, import the defined fleets and synchronise devices">
+                Resource syncs
+              </WithHelperText>
+            </CardTitle>
+            <CardBody>
+              <RepositoryResourceSyncList repositoryId={repositoryId} />
+            </CardBody>
+          </Card>
+        </GridItem>
+      </Grid>
       {deleteModal}
     </DetailsPage>
   );
